+++ examples/basics.js
@@ -1,28 +1,19 @@
-import pdfmake from "../dist/index.mjs"; // during development; use 'pdfmake' when installed
-import Roboto from "./fonts.js";
-pdfmake.addFonts(Roboto);
+import pdfcraft from "../dist/index.mjs"; // during development; use "pdfcraft" when installed
+import { configureExample } from "./setup.js";
+
+configureExample(pdfcraft);
 
 // or you can define the font manually:
 /*
-pdfmake.addFonts({
+pdfcraft.addFonts({
 	Roboto: {
-		normal: '../fonts/Roboto/Roboto-Regular.ttf',
-		bold: '../fonts/Roboto/Roboto-Medium.ttf',
-		italics: '../fonts/Roboto/Roboto-Italic.ttf',
-		bolditalics: '../fonts/Roboto/Roboto-MediumItalic.ttf'
+		normal: 'fonts/Roboto/Roboto-Regular.ttf',
+		bold: 'fonts/Roboto/Roboto-Medium.ttf',
+		italics: 'fonts/Roboto/Roboto-Italic.ttf',
+		bolditalics: 'fonts/Roboto/Roboto-MediumItalic.ttf'
 	}
 });
 */
-
-pdfmake.setUrlAccessPolicy((url) => {
-	// this can be used to restrict allowed domains
-	return url.startsWith("https://");
-});
-
-pdfmake.setLocalAccessPolicy((path) => {
-	// this can be used to restrict access to local file system
-	return true;
-});
 
 var docDefinition = {
 	content: [
@@ -33,7 +24,7 @@ var docDefinition = {
 
 var now = new Date();
 
-var pdf = pdfmake.createPdf(docDefinition);
+var pdf = pdfcraft.createPdf(docDefinition);
 pdf.write("pdfs/basics.pdf").then(
 	() => {
 		console.log(new Date() - now);